+++ etc/serialversions/serialversions.txt
@@ -4,13 +4,13 @@ org.woehlke.simpleworklist.domain.task.TaskState:    private static final long s
 org.woehlke.simpleworklist.domain.task.TaskTime:    private static final long serialVersionUID = 0L;
 org.woehlke.simpleworklist.user.resetpassword.UserPasswordRecoveryStatus:    private static final long serialVersionUID = 0L;
 org.woehlke.simpleworklist.user.register.UserRegistrationStatus:    private static final long serialVersionUID = 0L;
-org.woehlke.simpleworklist.user.account.UserRole:    private static final long serialVersionUID = 0L;
+org.woehlke.simpleworklist.user.domain.account.UserRole:    private static final long serialVersionUID = 0L;
 org.woehlke.simpleworklist.common.domain.AuditModel:    private static final long serialVersionUID = 4399373914714726911L;
 org.woehlke.simpleworklist.domain.context.Context:    private static final long serialVersionUID = -5035732370606951871L;
 org.woehlke.simpleworklist.domain.project.Project:    private static final long serialVersionUID = 4566653175832872422L;
 org.woehlke.simpleworklist.domain.task.Task:    private static final long serialVersionUID = 5247710652586269801L;
 org.woehlke.simpleworklist.user.chat.User2UserMessage:    private static final long serialVersionUID = 4263078228257938175L;
-org.woehlke.simpleworklist.user.account.UserAccount:    private static final long serialVersionUID = 7860692526488291439L;
+org.woehlke.simpleworklist.user.domain.account.UserAccount:    private static final long serialVersionUID = 7860692526488291439L;
 org.woehlke.simpleworklist.user.resetpassword.UserPasswordRecovery:    private static final long serialVersionUID = 6860716425733119940L;
 org.woehlke.simpleworklist.user.register.UserRegistration:    private static final long serialVersionUID = -1955967514018161878L;
 org.woehlke.simpleworklist.domain.breadcrumb.BreadcrumbItem:    private static final long serialVersionUID = 8276819198016077167L;
@@ -19,10 +19,10 @@ org.woehlke.simpleworklist.user.login.LoginForm:    private static final long se
 org.woehlke.simpleworklist.domain.context.NewContextForm:    private static final long serialVersionUID = -937143305653156981L;
 org.woehlke.simpleworklist.user.chat.User2UserMessageFormBean:    private static final long serialVersionUID = 1576610181966480168L;
 org.woehlke.simpleworklist.searchRequest.SearchResult:    private static final long serialVersionUID = 1682809351146047764L;
-org.woehlke.simpleworklist.user.account.UserAccountForm:    private static final long serialVersionUID = 9180383385243540190L;
-org.woehlke.simpleworklist.user.account.UserChangeDefaultContextForm:    private static final long serialVersionUID = -8592295563275083292L;
+org.woehlke.simpleworklist.user.domain.account.UserAccountForm:    private static final long serialVersionUID = 9180383385243540190L;
+org.woehlke.simpleworklist.user.domain.account.UserChangeDefaultContextForm:    private static final long serialVersionUID = -8592295563275083292L;
 org.woehlke.simpleworklist.common.language.UserChangeLanguageForm:    private static final long serialVersionUID = 2201123162578113187L;
-org.woehlke.simpleworklist.user.account.UserChangeNameForm:    private static final long serialVersionUID = 5120488382888268418L;
-org.woehlke.simpleworklist.user.account.UserChangePasswordForm:    private static final long serialVersionUID = 9149342594823222054L;
+org.woehlke.simpleworklist.user.domain.account.UserChangeNameForm:    private static final long serialVersionUID = 5120488382888268418L;
+org.woehlke.simpleworklist.user.domain.account.UserChangePasswordForm:    private static final long serialVersionUID = 9149342594823222054L;
 org.woehlke.simpleworklist.user.register.UserRegistrationForm:    private static final long serialVersionUID = 6864871862706880939L;
 org.woehlke.simpleworklist.user.session.UserSessionBean:    private static final long serialVersionUID = -6649686058228455825L;
