+++ src/main/java/org/woehlke/simpleworklist/common/testdata/TestDataService.java
@@ -1,6 +1,6 @@
 package org.woehlke.simpleworklist.common.testdata;
 
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 
 public interface TestDataService {
