+++ src/main/java/org/woehlke/simpleworklist/common/testdata/TestDataServiceImpl.java
@@ -9,7 +9,7 @@ import org.woehlke.simpleworklist.domain.project.Project;
 import org.woehlke.simpleworklist.domain.task.Task;
 import org.woehlke.simpleworklist.domain.task.TaskEnergy;
 import org.woehlke.simpleworklist.domain.task.TaskState;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 import org.woehlke.simpleworklist.domain.task.TaskTime;
 import org.woehlke.simpleworklist.domain.context.ContextRepository;
 import org.woehlke.simpleworklist.domain.task.TaskRepository;
