+++ src/main/java/org/woehlke/simpleworklist/domain/project/ProjectIdController.java
@@ -20,7 +20,7 @@ import org.woehlke.simpleworklist.domain.task.TaskState;
 import org.woehlke.simpleworklist.domain.task.TaskTime;
 import org.woehlke.simpleworklist.services.TaskProjektService;
 import org.woehlke.simpleworklist.domain.task.TaskService;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 import org.woehlke.simpleworklist.domain.breadcrumb.Breadcrumb;
 import org.woehlke.simpleworklist.user.session.UserSessionBean;
