+++ src/main/java/org/woehlke/simpleworklist/services/ProjectControllerService.java
@@ -6,7 +6,7 @@ import org.woehlke.simpleworklist.domain.context.Context;
 import org.woehlke.simpleworklist.domain.project.Project;
 import org.woehlke.simpleworklist.domain.task.Task;
 import org.woehlke.simpleworklist.user.session.UserSessionBean;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 
 import java.util.Locale;
