+++ src/main/java/org/woehlke/simpleworklist/services/ProjectControllerServiceImpl.java
@@ -15,7 +15,7 @@ import org.woehlke.simpleworklist.domain.project.ProjectService;
 import org.woehlke.simpleworklist.domain.task.Task;
 import org.woehlke.simpleworklist.domain.task.TaskService;
 import org.woehlke.simpleworklist.user.session.UserSessionBean;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
