+++ src/main/java/org/woehlke/simpleworklist/user/chat/User2UserMessage.java
@@ -5,7 +5,7 @@ import org.hibernate.validator.constraints.Length;
 import javax.validation.constraints.NotBlank;
 //import org.hibernate.validator.constraints.SafeHtml;
 import org.woehlke.simpleworklist.common.domain.AuditModel;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
