+++ src/main/java/org/woehlke/simpleworklist/user/chat/User2UserMessageController.java
@@ -17,7 +17,7 @@ import org.woehlke.simpleworklist.common.domain.AbstractController;
 import org.woehlke.simpleworklist.domain.breadcrumb.Breadcrumb;
 import org.woehlke.simpleworklist.user.session.UserSessionBean;
 import org.woehlke.simpleworklist.domain.context.Context;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 
 import javax.validation.Valid;
 import java.util.Locale;
