+++ src/main/java/org/woehlke/simpleworklist/user/domain/account/UserAccount.java
@@ -1,4 +1,4 @@
-package org.woehlke.simpleworklist.user.account;
+package org.woehlke.simpleworklist.user.domain.account;
 
 import javax.validation.constraints.Email;
