+++ src/main/java/org/woehlke/simpleworklist/user/domain/account/UserAccountForm.java
@@ -1,31 +1,38 @@
-package org.woehlke.simpleworklist.user.account;
+package org.woehlke.simpleworklist.user.domain.account;
 
+import javax.persistence.Transient;
+import javax.validation.constraints.NotNull;
+
+import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 
 import lombok.*;
 //import org.hibernate.validator.constraints.SafeHtml;
 
-import javax.persistence.Transient;
-import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 
-/**
- * Created by tw on 15.03.16.
- */
+
 @Getter
 @Setter
 @EqualsAndHashCode
 @AllArgsConstructor
 @NoArgsConstructor
-public class UserChangePasswordForm implements Serializable {
+@ToString(exclude = {"userPassword"})
+public class UserAccountForm implements Serializable {
 
-    private static final long serialVersionUID = 9149342594823222054L;
+    private static final long serialVersionUID = 9180383385243540190L;
+
+    //TODO: Messages i18n
+    @NotNull(message = "Email Address is compulsory")
+    @NotBlank(message = "Email Address is compulsory")
+    @Email(message = "Email Address is not a valid format")
+    private String userEmail;
 
     //TODO: Messages i18n
     //@SafeHtml(whitelistType= SafeHtml.WhiteListType.NONE)
-    @NotNull(message = "Password is compulsory")
-    @NotBlank(message = "Password is compulsory")
-    private String oldUserPassword;
+    @NotNull(message = "Fullname is compulsory")
+    @NotBlank(message = "Fullname is compulsory")
+    private String userFullname;
 
     //TODO: Messages i18n
     //@SafeHtml(whitelistType= SafeHtml.WhiteListType.NONE)
