+++ src/main/java/org/woehlke/simpleworklist/user/domain/account/UserAccountPassword.java
@@ -1,4 +1,4 @@
-package org.woehlke.simpleworklist.user.account;
+package org.woehlke.simpleworklist.user.domain.account;
 
 import lombok.Getter;
 import lombok.Setter;
