+++ src/main/java/org/woehlke/simpleworklist/user/domain/account/UserChangeDefaultContextForm.java
@@ -1,4 +1,4 @@
-package org.woehlke.simpleworklist.user.account;
+package org.woehlke.simpleworklist.user.domain.account;
 
 import lombok.*;
 import org.woehlke.simpleworklist.domain.context.Context;
