+++ src/main/java/org/woehlke/simpleworklist/user/register/UserRegistrationController.java
@@ -9,7 +9,7 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.woehlke.simpleworklist.user.account.UserAccountForm;
+import org.woehlke.simpleworklist.user.domain.account.UserAccountForm;
 import org.woehlke.simpleworklist.user.services.UserAccountService;
 import org.woehlke.simpleworklist.user.services.UserRegistrationService;
