+++ src/main/java/org/woehlke/simpleworklist/user/resetpassword/UserPasswordRecoveryController.java
@@ -8,8 +8,8 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.woehlke.simpleworklist.user.account.UserAccountForm;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccountForm;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 import org.woehlke.simpleworklist.user.register.UserRegistrationForm;
 import org.woehlke.simpleworklist.user.services.UserAccountService;
