+++ src/main/java/org/woehlke/simpleworklist/user/services/User2UserMessageService.java
@@ -2,7 +2,7 @@ package org.woehlke.simpleworklist.user.services;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 import org.woehlke.simpleworklist.user.chat.User2UserMessage;
 import org.woehlke.simpleworklist.user.chat.User2UserMessageFormBean;
