+++ src/main/java/org/woehlke/simpleworklist/user/services/User2UserMessageServiceImpl.java
@@ -6,13 +6,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.woehlke.simpleworklist.user.chat.User2UserMessage;
 import org.woehlke.simpleworklist.user.chat.User2UserMessageFormBean;
 import org.woehlke.simpleworklist.user.chat.User2UserMessageRepository;
-import org.woehlke.simpleworklist.user.services.User2UserMessageService;
 
 import java.util.List;
 import java.util.UUID;
