+++ src/main/java/org/woehlke/simpleworklist/user/services/UserAccountSecurityPasswordServiceImpl.java
@@ -11,10 +11,9 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.woehlke.simpleworklist.user.account.UserAccount;
-import org.woehlke.simpleworklist.user.account.UserDetailsBean;
-import org.woehlke.simpleworklist.user.account.UserAccountRepository;
-import org.woehlke.simpleworklist.user.services.UserAccountSecurityPasswordService;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserDetailsBean;
+import org.woehlke.simpleworklist.user.domain.account.UserAccountRepository;
 
 @Slf4j
 @Service
