+++ src/test/java/org/woehlke/simpleworklist/application/config/UserAccountTestDataService.java
@@ -1,6 +1,6 @@
 package org.woehlke.simpleworklist.application.config;
 
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 
 public interface UserAccountTestDataService {
