+++ src/test/java/org/woehlke/simpleworklist/application/config/UserAccountTestDataServiceImpl.java
@@ -9,7 +9,7 @@ import org.woehlke.simpleworklist.domain.context.Context;
 import org.woehlke.simpleworklist.domain.context.ContextService;
 import org.woehlke.simpleworklist.domain.context.NewContextForm;
 import org.woehlke.simpleworklist.common.language.Language;
-import org.woehlke.simpleworklist.user.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
 import org.woehlke.simpleworklist.user.services.UserAccountService;
 
 import java.util.Date;
