+++ src/test/java/org/woehlke/simpleworklist/user/UserDetailsBeanTest.java
@@ -3,8 +3,8 @@ package org.woehlke.simpleworklist.user;
 
 
 import org.springframework.security.core.GrantedAuthority;
-import org.woehlke.simpleworklist.user.account.UserAccount;
-import org.woehlke.simpleworklist.user.account.UserDetailsBean;
+import org.woehlke.simpleworklist.user.domain.account.UserAccount;
+import org.woehlke.simpleworklist.user.domain.account.UserDetailsBean;
 
 import java.util.Collection;
