--- src/test/java/org/woehlke/simpleworklist/user/account/UserAccountAccessServiceTest.java
@@ -1,4 +1,0 @@
-package org.woehlke.simpleworklist.user.account;
-
-public class UserAccountAccessServiceTest {
-}
