--- src/test/java/org/woehlke/simpleworklist/user/account/UserAccountServiceTest.java
@@ -1,4 +1,0 @@
-package org.woehlke.simpleworklist.user.account;
-
-public class UserAccountServiceTest {
-}
