+++ src/test/java/org/woehlke/simpleworklist/user/domain/account/UserAccountPasswordEncodedTest.java
@@ -1,4 +1,4 @@
-package org.woehlke.simpleworklist.user.account;
+package org.woehlke.simpleworklist.user.domain.account;
 
 
 import org.springframework.security.crypto.password.PasswordEncoder;
