+++ src/test/java/org/woehlke/simpleworklist/user/domain/account/UserAccountSecurityServiceTest.java
@@ -1,0 +1,4 @@
+package org.woehlke.simpleworklist.user.domain.account;
+
+public class UserAccountSecurityServiceTest {
+}
