+++ src/test/java/org/woehlke/simpleworklist/user/domain/account/UserAccountServiceImplTest.java
@@ -1,4 +1,4 @@
-package org.woehlke.simpleworklist.user.account;
+package org.woehlke.simpleworklist.user.domain.account;
 
 import org.springframework.beans.factory.annotation.Autowired;
